replace MCRMetadataManager with mock
